+++ packages/crypto/src/transactions/serializer.ts
@@ -7,7 +7,8 @@ import { Address } from "../identities";
 import { ITransaction, ITransactionData } from "../interfaces";
 import { ISerializeOptions } from "../interfaces";
 import { configManager } from "../managers/config";
-import { Base58, isSupportedTansactionVersion } from "../utils";
+import { isSupportedTansactionVersion } from "../utils";
+import { Base58 } from "../utils/base58";
 import { TransactionTypeFactory } from "./types";
 
 // Reference: https://github.com/ArkEcosystem/AIPs/blob/master/AIPS/aip-11.md
